requirements.txt
fastapi
uvicorn
langchain
openai
gradio
weaviate
pydub
simpleaudio
base64
huggingface-hub
transformers
python-dotenv